refactor: 클래스 용도별 패키지 생성 및 이동
