+++ src/main/java/team7/hrbank/domain/Department/repository/DepartmentRepository.java
@@ -1,4 +1,4 @@
-package team7.hrbank.domain.Department;
+package team7.hrbank.domain.Department.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import team7.hrbank.domain.Department.entity.Department;
 
 public interface DepartmentRepository extends JpaRepository<Department, Long> {
